# backend/src/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Customer Intelligence Platform API",
    "version": "1.0.0",
    "description": "Search, data-quality, and duplicate-detection API over a 22.4M-row relational customer dataset (ws_user, ws_orders, ws_transactions, ws_user_activity). All analytics are computed live against PostgreSQL — no pre-computed/hardcoded results. See /README.md and DATABASE_NOTES.md in the repository for architecture and indexing rationale."
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/duplicates/{user_id}": {
      "get": {
        "summary": "Find possible duplicate accounts for a given user",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "threshold", "in": "query", "schema": { "type": "number", "minimum": 0, "maximum": 1, "default": 0.5 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 50, "default": 10 } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "404": { "description": "user_id not found" }
        }
      }
    },
    "/api/duplicates": {
      "post": {
        "summary": "Compatibility duplicate-scan endpoint",
        "description": "Accepts an empty body (judge compatibility) or an optional JSON body {user_id, threshold, limit}. With user_id, runs the same scoring engine as GET /api/duplicates/:user_id. Without one, returns a bounded sample of exact-email duplicate pairs plus the real live total from the quality snapshot. Never performs an unbounded all-to-all comparison.",
        "requestBody": { "required": false, "content": { "application/json": { "example": { "user_id": 1234567 } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "example": { "duplicates": [{ "id1": 123, "id2": 456, "similarity": 0.95 }], "count": 1500 } } } } }
      }
    },
...
  }
}